Add `MultipleKeysTrait` to handle getMultiple/setMultiple/deleteMultiple
