+++ Client/MemoryClient.php
@@ -7,6 +7,8 @@ use Psr\SimpleCache\CacheInterface;
 
 class MemoryClient implements CacheInterface
 {
+    use MultipleKeysTrait;
+
     protected $cache   = [];
     protected $expires = [];
 
@@ -99,75 +101,6 @@ class MemoryClient implements CacheInterface
     {
         $this->cache   = [];
         $this->expires = [];
-
-        return true;
-    }
-
-    /**
-     * Obtains multiple cache items by their unique keys.
-     *
-     * @param iterable $keys    A list of keys that can obtained in a single operation.
-     * @param mixed    $default Default value to return for keys that do not exist.
-     *
-     * @return iterable A list of key => value pairs. Cache keys that do not exist or are stale will have $default as
-     *                  value.
-     *
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     *   MUST be thrown if $keys is neither an array nor a Traversable,
-     *   or if any of the $keys are not a legal value.
-     */
-    public function getMultiple($keys, $default = null)
-    {
-        return array_map(
-            function ($key) use ($default) {
-                return $this->get($key, $default);
-            },
-            array_combine((array)$keys, (array)$keys)
-        );
-    }
-
-    /**
-     * Persists a set of key => value pairs in the cache, with an optional TTL.
-     *
-     * @param iterable               $values A list of key => value pairs for a multiple-set operation.
-     * @param null|int|\DateInterval $ttl    Optional. The TTL value of this item. If no value is sent and
-     *                                       the driver supports TTL then the library may set a default value
-     *                                       for it or let the driver take care of that.
-     *
-     * @return bool True on success and false on failure.
-     *
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     *   MUST be thrown if $values is neither an array nor a Traversable,
-     *   or if any of the $values are not a legal value.
-     */
-    public function setMultiple($values, $ttl = null)
-    {
-        foreach ($values as $key => $value) {
-            $this->set($key, $value, $ttl);
-        }
-
-        return true;
-    }
-
-    /**
-     * Deletes multiple cache items in a single operation.
-     *
-     * @param iterable $keys A list of string-based keys to be deleted.
-     *
-     * @return bool True if the items were successfully removed. False if there was an error.
-     *
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     *   MUST be thrown if $keys is neither an array nor a Traversable,
-     *   or if any of the $keys are not a legal value.
-     */
-    public function deleteMultiple($keys)
-    {
-        array_map(
-            function ($key) {
-                return $this->delete($key);
-            },
-            (array)$keys
-        );
 
         return true;
     }
